+++ musicalsunburst/music/apps.py
@@ -1,4 +1,5 @@
 from django.apps import AppConfig
 
+
 class MusicConfig(AppConfig):
-    name='music'
+    name = 'music'
