+++ musicalsunburst/musicalsunburst/urls.py
@@ -17,6 +17,7 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
+    path('music/', include('music.urls')),
     path('sunburst/', include('sunburst.urls')),
     path('admin/', admin.site.urls),
 ]
