--- website/music/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from .models import Album ,Song
-
-admin.site.register(Album)
-admin.site.register(Song)
