--- website/music/forms.py
@@ -1,9 +1,0 @@
-from django.contrib.auth.models import User
-from django import forms
-
-class UserForm(forms.ModelForm):
-    password = forms.CharField(widget=forms.PasswordInput)
-
-    class Meta:
-        model=User
-        Fields = ['Name' , 'Username' , 'email' , 'password']
